data_pipelines_with_airflow/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
apache-airflow==2.10.5
apache-airflow-providers-amazon==9.2.0
apache-airflow-providers-common-compat==1.3.0
apache-airflow-providers-common-io==1.5.0
apache-airflow-providers-common-sql==1.21.0
apache-airflow-providers-fab==1.5.3
apache-airflow-providers-ftp==3.12.0
apache-airflow-providers-google==12.0.0
apache-airflow-providers-http==5.0.0
apache-airflow-providers-imap==3.8.0
apache-airflow-providers-postgres==6.0.0
apache-airflow-providers-smtp==1.9.0
apache-airflow-providers-sqlite==4.0.0
apispec==6.8.1
argcomplete==3.5.3
asgiref==3.8.1
asn1crypto==1.5.1
asyncpg==0.30.0
attrs==25.1.0
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.13.3
blinker==1.9.0
boto3==1.36.19
botocore==1.36.19
cachelib==0.9.0
cachetools==5.5.1
cattrs==24.1.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
clickclick==20.10.2
colorama==0.4.6
colorlog==6.9.0
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==44.0.1
db-dtypes==1.4.1
decorator==5.1.1
Deprecated==1.2.18
dill==0.3.9
dnspython==2.7.0
docstring_parser==0.16
durationpy==0.9
email_validator==2.2.0
Flask==2.2.5
Flask-AppBuilder==4.5.3
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.10.1
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
frozenlist==1.5.0
fsspec==2025.2.0
gcloud-aio-auth==5.3.2
gcloud-aio-bigquery==7.1.0
gcloud-aio-storage==9.3.0
gcsfs==2025.2.0
google-ads==25.1.0
google-analytics-admin==0.23.4
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.80.0
google-cloud-alloydb==0.4.2
google-cloud-appengine-logging==1.6.0
google-cloud-audit-log==0.3.0
google-cloud-automl==2.16.0
google-cloud-batch==0.17.34
google-cloud-bigquery==3.29.0
google-cloud-bigquery-datatransfer==3.19.0
google-cloud-bigtable==2.28.1
google-cloud-build==3.30.0
google-cloud-compute==1.25.0
google-cloud-container==2.56.0
google-cloud-core==2.4.1
google-cloud-datacatalog==3.25.0
google-cloud-dataflow-client==0.8.16
google-cloud-dataform==0.5.15
google-cloud-dataplex==2.7.0
google-cloud-dataproc==5.17.0
google-cloud-dataproc-metastore==1.18.0
google-cloud-dlp==3.27.0
google-cloud-kms==3.3.0
google-cloud-language==2.17.0
google-cloud-logging==3.11.4
google-cloud-memcache==1.12.0
google-cloud-monitoring==2.27.0
google-cloud-orchestration-airflow==1.17.0
google-cloud-os-login==2.17.0
google-cloud-pubsub==2.28.0
google-cloud-redis==2.17.0
google-cloud-resource-manager==1.14.0
google-cloud-run==0.10.15
google-cloud-secret-manager==2.23.0
google-cloud-spanner==3.51.0
google-cloud-speech==2.31.0
google-cloud-storage==2.19.0
google-cloud-storage-transfer==1.16.0
google-cloud-tasks==2.19.0
google-cloud-texttospeech==2.25.0
google-cloud-translate==3.20.0
google-cloud-videointelligence==2.16.0
google-cloud-vision==3.10.0
google-cloud-workflows==1.17.0
google-crc32c==1.6.0
google-re2==1.1.20240702
google-resumable-media==2.7.2
googleapis-common-protos==1.67.0
greenlet==3.1.1
grpc-google-iam-v1==0.14.0
grpc-interceptor==0.15.4
grpcio==1.70.0
grpcio-gcp==0.2.2
grpcio-status==1.70.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
immutabledict==4.2.1
importlib_metadata==8.5.0
inflection==0.5.1
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
json-merge-patch==0.2
jsonpath-ng==1.7.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==32.0.0
lazy-object-proxy==1.10.0
limits==4.0.1
linkify-it-py==2.0.3
lockfile==0.12.2
looker-sdk==25.2.0
lxml==5.3.1
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.6.0
multidict==6.1.0
numpy==1.26.4
oauthlib==3.2.2
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
ordered-set==4.1.0
packaging==24.2
pandas==2.1.4
pandas-gbq==0.27.0
pathspec==0.12.1
pendulum==3.0.0
pluggy==1.5.0
ply==3.11
prison==0.2.1
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
psutil==6.1.1
psycopg2-binary==2.9.10
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.0
PyAthena==3.12.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydata-google-auth==1.9.1
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-nvd3==0.16.0
python-slugify==8.0.4
python3-saml==1.16.0
pytz==2025.1
PyYAML==6.0.2
redshift-connector==2.1.5
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.7.0
rpds-py==0.22.3
rsa==4.9
s3transfer==0.11.2
scramp==1.4.5
setproctitle==1.3.4
setuptools==75.8.0
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==1.4.54
sqlalchemy-bigquery==1.12.1
SQLAlchemy-JSONField==1.0.2
sqlalchemy-spanner==1.8.0
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.5.0
text-unidecode==1.3
time-machine==2.16.0
typing_extensions==4.12.2
tzdata==2025.1
uc-micro-py==1.0.3
universal_pathlib==0.2.6
uritemplate==4.1.1
urllib3==2.3.0
watchtower==3.3.1
websocket-client==1.8.0
Werkzeug==2.2.3
wirerope==1.0.0
wrapt==1.17.2
WTForms==3.2.1
xmlsec==1.3.14
yarl==1.18.3
zipp==3.21.0
